feat: DependencyCheckTaskConfig extracted from AnalyzeTaskConfig for Purge and Update tasks
